Ultimate Predictive Model
Based of the prompt in the challenge doc, I created a predictive model to determine whether or not a user will be active into the future. Additionally, I extracted key indicators that put a user at risk for becoming inactive as well as the importance of each feature in determining user activity.